requirements/common.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements/common.in
#
aiohappyeyeballs==2.4.4
    # via aiohttp
aiohttp==3.11.11
    # via -r requirements/common.in
aiosignal==1.3.2
    # via aiohttp
apscheduler==3.11.0
    # via django-apscheduler
asgiref==3.8.1
    # via django
attrs==24.3.0
    # via
    #   aiohttp
    #   jsonschema
    #   referencing
babel==2.16.0
    # via py-moneyed
build==1.2.2.post1
    # via pip-tools
certifi==2024.12.14
    # via
    #   django-money
    #   requests
cfgv==3.4.0
    # via pre-commit
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via pip-tools
coreapi==2.3.3
    # via
    #   -r requirements/common.in
    #   openapi-codec
coreschema==0.0.4
    # via
    #   -r requirements/common.in
    #   coreapi
coverage[toml]==7.6.10
    # via
    #   -r requirements/common.in
    #   pytest-cov
    #   pytest-testmon
debugpy==1.8.12
    # via -r requirements/common.in
diff-match-patch==20241021
    # via django-import-export
distlib==0.3.9
    # via virtualenv
django==5.1.5
    # via
    #   -r requirements/common.in
    #   django-admin-autocomplete-filter
    #   django-admin-numeric-filter
    #   django-admin-sortable2
    #   django-apscheduler
    #   django-debug-toolbar
    #   django-dramatiq
    #   django-extensions
    #   django-filter
    #   django-import-export
    #   django-money
    #   django-phonenumber-field
    #   django-unfold
    #   djangorestframework
    #   drf-spectacular
    #   drf-yasg
django-admin-autocomplete-all==0.6.6
    # via -r requirements/common.in
django-admin-autocomplete-filter==0.7.1
    # via -r requirements/common.in
django-admin-list-filter-dropdown==1.0.3
    # via -r requirements/common.in
django-admin-numeric-filter==0.1.9
    # via -r requirements/common.in
django-admin-rangefilter==0.13.2
    # via -r requirements/common.in
django-admin-sortable2==2.2.4
    # via -r requirements/common.in
django-advanced-filters==2.0.0
    # via -r requirements/common.in
django-apscheduler==0.7.0
    # via -r requirements/common.in
django-debug-toolbar==5.0.1
    # via -r requirements/common.in
django-dramatiq==0.13.0
    # via -r requirements/common.in
django-extensions==3.2.3
    # via -r requirements/common.in
django-filter==24.3
    # via -r requirements/common.in
django-import-export==4.3.4
    # via -r requirements/common.in
django-json-widget==2.0.1
    # via -r requirements/common.in
django-log-viewer==1.1.7
    # via -r requirements/common.in
django-macaddress==1.8.0
    # via -r requirements/common.in
django-money[exchange]==3.5.3
    # via -r requirements/common.in
django-phonenumber-field[phonenumbers]==8.0.0
    # via -r requirements/common.in
django-unfold==0.45.0
    # via -r requirements/common.in
djangorestframework==3.15.2
    # via
    #   -r requirements/common.in
    #   drf-spectacular
    #   drf-yasg
dramatiq[rabbitmq,watch]==1.17.1
    # via
    #   -r requirements/common.in
    #   django-dramatiq
drf-spectacular==0.28.0
    # via -r requirements/common.in
drf-yasg[validation]==1.21.8
    # via -r requirements/common.in
et-xmlfile==2.0.0
    # via openpyxl
execnet==2.1.1
    # via pytest-xdist
factory-boy==3.3.1
    # via
    #   -r requirements/common.in
    #   pytest-factoryboy
faker==35.0.0
    # via
    #   -r requirements/common.in
    #   factory-boy
filelock==3.17.0
    # via virtualenv
flake8==7.1.1
    # via -r requirements/common.in
fluent-logger==0.11.1
    # via -r requirements/common.in
frozenlist==1.5.0
    # via
    #   aiohttp
    #   aiosignal
gevent==24.11.1
    # via watchdog-gevent
greenlet==3.1.1
    # via gevent
gunicorn==23.0.0
    # via -r requirements/common.in
httptools==0.6.4
    # via -r requirements/common.in
identify==2.6.6
    # via pre-commit
idna==3.10
    # via
    #   requests
    #   yarl
importlib-resources==6.5.2
    # via swagger-spec-validator
inflection==0.5.1
    # via
    #   drf-spectacular
    #   drf-yasg
    #   pytest-factoryboy
iniconfig==2.0.0
    # via pytest
isodate==0.7.2
    # via -r requirements/common.in
itypes==1.2.0
    # via coreapi
jinja2==3.1.5
    # via coreschema
json-log-formatter==1.1
    # via -r requirements/common.in
jsonschema==4.23.0
    # via
    #   -r requirements/common.in
    #   drf-spectacular
    #   swagger-spec-validator
jsonschema-specifications==2024.10.1
    # via jsonschema
markupsafe==3.0.2
    # via jinja2
mccabe==0.7.0
    # via flake8
msgpack==1.1.0
    # via fluent-logger
multidict==6.1.0
    # via
    #   aiohttp
    #   yarl
netaddr==1.3.0
    # via django-macaddress
nodeenv==1.9.1
    # via pre-commit
openapi-codec==1.3.2
    # via -r requirements/common.in
openpyxl==3.1.5
    # via
    #   -r requirements/common.in
    #   tablib
packaging==24.2
    # via
    #   build
    #   drf-yasg
    #   gunicorn
    #   pytest
    #   pytest-factoryboy
phonenumbers==8.13.53
    # via django-phonenumber-field
pika==1.3.2
    # via dramatiq
pip-tools==7.4.1
    # via -r requirements/common.in
platformdirs==4.3.6
    # via virtualenv
pluggy==1.5.0
    # via pytest
pre-commit==4.1.0
    # via -r requirements/common.in
prometheus-client==0.21.1
    # via dramatiq
propcache==0.2.1
    # via
    #   aiohttp
    #   yarl
psycopg2-binary==2.9.10
    # via -r requirements/common.in
py-moneyed==3.0
    # via django-money
pycodestyle==2.12.1
    # via flake8
pyflakes==3.2.0
    # via flake8
pygraphviz==1.14
    # via -r requirements/common.in
pyproject-hooks==1.2.0
    # via
    #   build
    #   pip-tools
pytest==8.3.4
    # via
    #   -r requirements/common.in
    #   pytest-cov
    #   pytest-django
    #   pytest-factoryboy
    #   pytest-testmon
    #   pytest-xdist
pytest-cov==6.0.0
    # via -r requirements/common.in
pytest-django==4.9.0
    # via -r requirements/common.in
pytest-factoryboy==2.7.0
    # via -r requirements/common.in
pytest-testmon==2.1.3
    # via -r requirements/common.in
pytest-xdist==3.6.1
    # via -r requirements/common.in
python-dateutil==2.9.0.post0
    # via faker
pytz==2024.2
    # via drf-yasg
pyyaml==6.0.2
    # via
    #   drf-spectacular
    #   drf-yasg
    #   pre-commit
    #   swagger-spec-validator
referencing==0.36.1
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.3
    # via
    #   -r requirements/common.in
    #   coreapi
    #   requests-mock
requests-mock==1.12.1
    # via -r requirements/common.in
rpds-py==0.22.3
    # via
    #   jsonschema
    #   referencing
simplejson==3.19.3
    # via django-advanced-filters
six==1.17.0
    # via
    #   python-dateutil
    #   smpplib
smpplib==2.2.4
    # via -r requirements/common.in
snakeviz==2.2.2
    # via -r requirements/common.in
sqlparse==0.5.3
    # via
    #   django
    #   django-debug-toolbar
swagger-spec-validator==3.0.4
    # via drf-yasg
tablib[xls,xlsx]==3.8.0
    # via
    #   -r requirements/common.in
    #   django-import-export
tornado==6.4.2
    # via snakeviz
typing-extensions==4.12.2
    # via
    #   faker
    #   py-moneyed
    #   pytest-factoryboy
    #   referencing
    #   swagger-spec-validator
tzlocal==5.2
    # via apscheduler
uritemplate==4.1.1
    # via
    #   -r requirements/common.in
    #   coreapi
    #   drf-spectacular
    #   drf-yasg
urllib3==2.3.0
    # via
    #   -r requirements/common.in
    #   requests
virtualenv==20.29.1
    # via pre-commit
watchdog==6.0.0
    # via
    #   dramatiq
    #   watchdog-gevent
watchdog-gevent==0.2.1
    # via dramatiq
wheel==0.45.1
    # via pip-tools
xlrd==2.0.1
    # via tablib
xlwt==1.3.0
    # via tablib
yarl==1.18.3
    # via aiohttp
zope-event==5.0
    # via gevent
zope-interface==7.2
    # via gevent

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
